+++ manuscript/Changes.md
@@ -1,5 +1,7 @@
 # Changes
 
+* Add Hello world examples
+* Update calc.psgi, add calc.t
 
 ## 2017-09-26